refactor(react): use PascalCase for file and folder naming
