+++ packages/react/src/api/common/index.ts
@@ -1,2 +1,2 @@
-export * from "./use-mutation";
-export * from "./use-query";
+export * from "./UseMutation";
+export * from "./UseQuery";
