+++ packages/react/src/components/index.ts
@@ -1,4 +1,4 @@
 export * from "./account";
 export * from "./accounts";
-export * from "./my-accounts";
-export * from "./signup";
+export * from "./MyAccounts";
+export * from "./Signup";
