+++ packages/react/src/hooks/accounts/index.ts
@@ -1,6 +1,6 @@
-export * from "./use-add-account-mutation";
-export * from "./use-edit-account-mutation";
-export * from "./use-get-account-query";
-export * from "./use-get-my-account-query";
+export * from "./UseAddAccountMutation";
+export * from "./UseEditAccountMutation";
+export * from "./UseGetAccountQuery";
+export * from "./UseGetMyAccountQuery";
 export * from "./invitations";
 export * from "./users";
