+++ packages/react/src/hooks/accounts/invitations/index.ts
@@ -1,8 +1,8 @@
-export * from "./use-add-invitation-mutation";
-export * from "./use-delete-invitation-mutation";
-export * from "./use-get-invitation-query";
-export * from "./use-get-invitations-query";
-export * from "./use-join-invitation-mutation";
-export * from "./use-resend-invitation-mutation";
-export * from "./use-revoke-invitation-mutation";
-export * from "./use-signup-invitation-mutation";
+export * from "./UseAddInvitationMutation";
+export * from "./UseDeleteInvitationMutation";
+export * from "./UseGetInvitationQuery";
+export * from "./UseGetInvitationsQuery";
+export * from "./UseJoinInvitationMutation";
+export * from "./UseResendInvitationMutation";
+export * from "./UseRevokeInvitationMutation";
+export * from "./UseSignupInvitationMutation";
