+++ packages/react/src/hooks/accounts/users/index.ts
@@ -1,3 +1,3 @@
-export * from "./use-disable-user-mutation";
-export * from "./use-enable-user-mutation";
-export * from "./use-get-users-query";
+export * from "./UseDisableUserMutation";
+export * from "./UseEnableUserMutation";
+export * from "./UseGetUsersQuery";
