+++ packages/react/src/types/index.ts
@@ -1,6 +1,6 @@
 export * from "./account";
-export * from "./account-invitation";
-export * from "./account-user";
+export * from "./AccountInvitation";
+export * from "./AccountUser";
 export * from "./config";
 export * from "./routes";
 export * from "./types";
