+++ packages/react/src/views/AcceptInvitation/SignupInvitation.tsx
@@ -5,7 +5,7 @@ import { useParams, useSearchParams } from "react-router-dom";
 import { toast } from "react-toastify";
 
 import { UseMutationRequestObject } from "@/api";
-import { UserSignupForm } from "@/components/signup";
+import { UserSignupForm } from "@/components/Signup";
 import { useGetInvitationQuery, useSignupInvitationMutation } from "@/hooks";
 import { AcceptInvitationResponse, UserSignupData } from "@/types";
